fixed norme errors

+++ ft_memccpy.c
@@ -1,15 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memccpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgoyette <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/29 09:14:05 by jgoyette          #+#    #+#             */
+/*   Updated: 2018/11/29 09:23:56 by jgoyette         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*ptr_dest;
 	unsigned char	*ptr_src;
 
 	i = 0;
 	ptr_dest = (unsigned char *)dest;
 	ptr_src = (unsigned char *)src;
-	while ((size_t)i < n)
+	while (i < n)
 	{
 		ptr_dest[i] = ptr_src[i];
 		if (ptr_dest[i] == (unsigned int)c)
