+++ ft_memcmp.c
@@ -1,15 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcmp.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgoyette <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/29 09:14:26 by jgoyette          #+#    #+#             */
+/*   Updated: 2018/11/29 09:25:34 by jgoyette         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*ptr_s1;
 	unsigned char	*ptr_s2;
 
 	i = 0;
 	ptr_s1 = (unsigned char *)s1;
 	ptr_s2 = (unsigned char *)s2;
-	while ((size_t)i < n)
+	while (i < n)
 	{
 		if (ptr_s1[i] != ptr_s2[i])
 			return ((int)(ptr_s1[i] - ptr_s2[i]));
