+++ ft_memcpy.c
@@ -1,15 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgoyette <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/29 09:14:00 by jgoyette          #+#    #+#             */
+/*   Updated: 2018/11/29 09:18:23 by jgoyette         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int 		i;
+	size_t				i;
 	unsigned char		*ptr_dest;
 	const unsigned char	*ptr_src;
 
 	i = 0;
 	ptr_dest = (unsigned char *)dest;
 	ptr_src = (const unsigned char *)src;
-	while ((size_t)i < n)
+	while (i < n)
 	{
 		ptr_dest[i] = ptr_src[i];
 		i += 1;
