+++ ft_strlen.c
@@ -1,11 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlen.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgoyette <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/29 09:13:14 by jgoyette          #+#    #+#             */
+/*   Updated: 2018/11/29 09:13:35 by jgoyette         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	unsigned int count;
+	size_t count;
 
 	count = 0;
 	while (str[count])
 		count += 1;
-	return ((size_t)count);
+	return (count);
 }
