+++ ft_strncat.c
@@ -1,13 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strncat.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgoyette <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/11/29 09:12:24 by jgoyette          #+#    #+#             */
+/*   Updated: 2018/11/29 09:30:55 by jgoyette         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 char	*ft_strncat(char *dest, const char *src, size_t n)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
-	i = (unsigned int)ft_strlen(dest);
+	i = ft_strlen(dest);
 	j = 0;
-	while (src[j] && (size_t)j < n)
+	while (src[j] && j < n)
 	{
 		dest[i] = src[j];
 		i += 1;
